+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- `build` shell command for python.
+
 ### Fixed
 - Terraform deploy will now exit with non-zero exit code if deploy failed.
 - Give pylint a $HOME so it won't complain when writing cache and stats.